+++ src/Umbraco.Web.UI.Client/src/packages/data-type/repository/detail/data-type-detail.repository.ts
@@ -1,5 +1,5 @@
 import type { UmbDataTypeDetailModel } from '../../types.js';
-import { UmbDataTypeServerDataSource } from './data-type-detail.server.data-source.js';
+import { UmbDataTypeServerDataSource } from './server-data-source/data-type-detail.server.data-source.js';
 import type { UmbDataTypeDetailStore } from './data-type-detail.store.js';
 import { UMB_DATA_TYPE_DETAIL_STORE_CONTEXT } from './data-type-detail.store.context-token.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
